+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/openid/PresentationFactory.kt
@@ -143,9 +143,11 @@ internal class PresentationFactory(
             docType = docType,
             namespaces = deviceNameSpaceBytes
         )
-        val deviceAuthenticationBytes =
-            coseCompliantSerializer.encodeToByteArray(ByteStringWrapper(deviceAuthentication)).wrapInCborTag(24)
-        Napier.i("Device authentication is ${deviceAuthenticationBytes.encodeToString(Base16())}")
+        val deviceAuthenticationBytes = coseCompliantSerializer
+            .encodeToByteArray(ByteStringWrapper(deviceAuthentication))
+            .wrapInCborTag(24)
+            .also { Napier.d("Device authentication signature input is ${it.encodeToString(Base16())}") }
+
         coseService.createSignedCoseWithDetachedPayload(
             payload = deviceAuthenticationBytes,
             serializer = ByteArraySerializer(),